@@ -7,7 +7,6 @@ using System.Collections.Generic;
 using System.Globalization;
 using System.IO;
 using System.Linq;
-using System.Runtime.InteropServices;
 using System.Text.RegularExpressions;
 using System.Threading;
 using System.Threading.Tasks;
@@ -68,7 +67,7 @@ namespace Microsoft.Diagnostics.NETCore.Client
         /// </returns> 
         public EventPipeSession StartEventPipeSession(IEnumerable<EventPipeProvider> providers, bool requestRundown = true, int circularBufferMB = 256)
         {
-            return new EventPipeSession(_endpoint, providers, requestRundown, circularBufferMB);
+            return EventPipeSession.Start(_endpoint, providers, requestRundown, circularBufferMB);
         }
 
         /// <summary>
@@ -82,7 +81,37 @@ namespace Microsoft.Diagnostics.NETCore.Client
         /// </returns> 
         public EventPipeSession StartEventPipeSession(EventPipeProvider provider, bool requestRundown = true, int circularBufferMB = 256)
         {
-            return new EventPipeSession(_endpoint, new[] { provider }, requestRundown, circularBufferMB);
+            return EventPipeSession.Start(_endpoint, new[] { provider }, requestRundown, circularBufferMB);
+        }
+
+        /// <summary>
+        /// Start tracing the application and return an EventPipeSession object
+        /// </summary>
+        /// <param name="providers">An IEnumerable containing the list of Providers to turn on.</param>
+        /// <param name="requestRundown">If true, request rundown events from the runtime</param>
+        /// <param name="circularBufferMB">The size of the runtime's buffer for collecting events in MB</param>
+        /// <param name="token">The token to monitor for cancellation requests.</param>
+        /// <returns>
+        /// An EventPipeSession object representing the EventPipe session that just started.
+        /// </returns> 
+        internal Task<EventPipeSession> StartEventPipeSessionAsync(IEnumerable<EventPipeProvider> providers, bool requestRundown, int circularBufferMB, CancellationToken token)
+        {
+            return EventPipeSession.StartAsync(_endpoint, providers, requestRundown, circularBufferMB, token);
+        }
+
+        /// <summary>
+        /// Start tracing the application and return an EventPipeSession object
+        /// </summary>
+        /// <param name="provider">An EventPipeProvider to turn on.</param>
+        /// <param name="requestRundown">If true, request rundown events from the runtime</param>
+        /// <param name="circularBufferMB">The size of the runtime's buffer for collecting events in MB</param>
+        /// <param name="token">The token to monitor for cancellation requests.</param>
+        /// <returns>
+        /// An EventPipeSession object representing the EventPipe session that just started.
+        /// </returns>
+        internal Task<EventPipeSession> StartEventPipeSessionAsync(EventPipeProvider provider, bool requestRundown, int circularBufferMB, CancellationToken token)
+        {
+            return EventPipeSession.StartAsync(_endpoint, new[] { provider }, requestRundown, circularBufferMB, token);
         }
 
         /// <summary>
@@ -93,26 +122,23 @@ namespace Microsoft.Diagnostics.NETCore.Client
         /// <param name="logDumpGeneration">When set to true, display the dump generation debug log to the console.</param>
         public void WriteDump(DumpType dumpType, string dumpPath, bool logDumpGeneration = false)
         {
-            if (string.IsNullOrEmpty(dumpPath))
-                throw new ArgumentNullException($"{nameof(dumpPath)} required");
+            IpcMessage request = CreateWriteDumpMessage(dumpType, dumpPath, logDumpGeneration);
+            IpcMessage response = IpcClient.SendMessage(_endpoint, request);
+            ValidateResponseMessage(response, nameof(WriteDump));
+        }
 
-            byte[] payload = SerializePayload(dumpPath, (uint)dumpType, logDumpGeneration);
-            IpcMessage message = new IpcMessage(DiagnosticsServerCommandSet.Dump, (byte)DumpCommandId.GenerateCoreDump, payload);
-            IpcMessage response = IpcClient.SendMessage(_endpoint, message);
-            switch ((DiagnosticsServerResponseId)response.Header.CommandId)
-            {
-                case DiagnosticsServerResponseId.Error:
-                    uint hr = BitConverter.ToUInt32(response.Payload, 0);
-                    if (hr == (uint)DiagnosticsIpcError.UnknownCommand)
-                    {
-                        throw new UnsupportedCommandException($"Unsupported operating system: {RuntimeInformation.OSDescription}");
-                    }
-                    throw new ServerErrorException($"Writing dump failed (HRESULT: 0x{hr:X8})");
-                case DiagnosticsServerResponseId.OK:
-                    return;
-                default:
-                    throw new ServerErrorException($"Writing dump failed - server responded with unknown command");
-            }
+        /// <summary>
+        /// Trigger a core dump generation.
+        /// </summary> 
+        /// <param name="dumpType">Type of the dump to be generated</param>
+        /// <param name="dumpPath">Full path to the dump to be generated. By default it is /tmp/coredump.{pid}</param>
+        /// <param name="logDumpGeneration">When set to true, display the dump generation debug log to the console.</param>
+        /// <param name="token">The token to monitor for cancellation requests.</param>
+        internal async Task WriteDumpAsync(DumpType dumpType, string dumpPath, bool logDumpGeneration, CancellationToken token)
+        {
+            IpcMessage request = CreateWriteDumpMessage(dumpType, dumpPath, logDumpGeneration);
+            IpcMessage response = await IpcClient.SendMessageAsync(_endpoint, request, token).ConfigureAwait(false);
+            ValidateResponseMessage(response, nameof(WriteDumpAsync));
         }
 
         /// <summary>
@@ -124,41 +150,20 @@ namespace Microsoft.Diagnostics.NETCore.Client
         /// <param name="additionalData">Additional data to be passed to the profiler</param>
         public void AttachProfiler(TimeSpan attachTimeout, Guid profilerGuid, string profilerPath, byte[] additionalData = null)
         {
-            if (profilerGuid == null || profilerGuid == Guid.Empty)
-            {
-                throw new ArgumentException($"{nameof(profilerGuid)} must be a valid Guid");
-            }
-
-            if (String.IsNullOrEmpty(profilerPath))
-            {
-                throw new ArgumentException($"{nameof(profilerPath)} must be non-null");
-            }
-
-            byte[] serializedConfiguration = SerializePayload((uint)attachTimeout.TotalSeconds, profilerGuid, profilerPath, additionalData);
-            var message = new IpcMessage(DiagnosticsServerCommandSet.Profiler, (byte)ProfilerCommandId.AttachProfiler, serializedConfiguration);
-            var response = IpcClient.SendMessage(_endpoint, message);
-            switch ((DiagnosticsServerResponseId)response.Header.CommandId)
-            {
-                case DiagnosticsServerResponseId.Error:
-                    uint hr = BitConverter.ToUInt32(response.Payload, 0);
-                    if (hr == (uint)DiagnosticsIpcError.UnknownCommand)
-                    {
-                        throw new UnsupportedCommandException("The target runtime does not support profiler attach");
-                    }
-                    if (hr == (uint)DiagnosticsIpcError.ProfilerAlreadyActive)
-                    {
-                        throw new ProfilerAlreadyActiveException("The request to attach a profiler was denied because a profiler is already loaded");
-                    }
-                    throw new ServerErrorException($"Profiler attach failed (HRESULT: 0x{hr:X8})");
-                case DiagnosticsServerResponseId.OK:
-                    return;
-                default:
-                    throw new ServerErrorException($"Profiler attach failed - server responded with unknown command");
-            }
+            IpcMessage request = CreateAttachProfilerMessage(attachTimeout, profilerGuid, profilerPath, additionalData);
+            IpcMessage response = IpcClient.SendMessage(_endpoint, request);
+            ValidateResponseMessage(response, nameof(AttachProfiler));
 
             // The call to set up the pipe and send the message operates on a different timeout than attachTimeout, which is for the runtime.
             // We should eventually have a configurable timeout for the message passing, potentially either separately from the 
             // runtime timeout or respect attachTimeout as one total duration.
+        }
+
+        internal async Task AttachProfilerAsync(TimeSpan attachTimeout, Guid profilerGuid, string profilerPath, byte[] additionalData, CancellationToken token)
+        {
+            IpcMessage request = CreateAttachProfilerMessage(attachTimeout, profilerGuid, profilerPath, additionalData);
+            IpcMessage response = await IpcClient.SendMessageAsync(_endpoint, request, token).ConfigureAwait(false);
+            ValidateResponseMessage(response, nameof(AttachProfilerAsync));
         }
 
         /// <summary>
@@ -169,38 +174,16 @@ namespace Microsoft.Diagnostics.NETCore.Client
         /// <param name="profilerPath">Path to the profiler to be attached</param>
         public void SetStartupProfiler(Guid profilerGuid, string profilerPath)
         {
-            if (profilerGuid == null || profilerGuid == Guid.Empty)
-            {
-                throw new ArgumentException($"{nameof(profilerGuid)} must be a valid Guid");
-            }
+            IpcMessage request = CreateSetStartupProfilerMessage(profilerGuid, profilerPath);
+            IpcMessage response = IpcClient.SendMessage(_endpoint, request);
+            ValidateResponseMessage(response, nameof(SetStartupProfiler), ValidateResponseOptions.InvalidArgumentIsRequiresSuspension);
+        }
 
-            if (String.IsNullOrEmpty(profilerPath))
-            {
-                throw new ArgumentException($"{nameof(profilerPath)} must be non-null");
-            }
-
-            byte[] serializedConfiguration = SerializePayload(profilerGuid, profilerPath);
-            var message = new IpcMessage(DiagnosticsServerCommandSet.Profiler, (byte)ProfilerCommandId.StartupProfiler, serializedConfiguration);
-            var response = IpcClient.SendMessage(_endpoint, message);
-            switch ((DiagnosticsServerResponseId)response.Header.CommandId)
-            {
-                case DiagnosticsServerResponseId.Error:
-                    uint hr = BitConverter.ToUInt32(response.Payload, 0);
-                    if (hr == (uint)DiagnosticsIpcError.UnknownCommand)
-                    {
-                        throw new UnsupportedCommandException("The target runtime does not support the ProfilerStartup command.");
-                    }
-                    else if (hr == (uint)DiagnosticsIpcError.InvalidArgument)
-                    {
-                        throw new ServerErrorException("The runtime must be suspended to issue the SetStartupProfiler command.");
-                    }
-
-                    throw new ServerErrorException($"Profiler startup failed (HRESULT: 0x{hr:X8})");
-                case DiagnosticsServerResponseId.OK:
-                    return;
-                default:
-                    throw new ServerErrorException($"Profiler startup failed - server responded with unknown command");
-            }
+        internal async Task SetStartupProfilerAsync(Guid profilerGuid, string profilerPath, CancellationToken token)
+        {
+            IpcMessage request = CreateSetStartupProfilerMessage(profilerGuid, profilerPath);
+            IpcMessage response = await IpcClient.SendMessageAsync(_endpoint, request, token).ConfigureAwait(false);
+            ValidateResponseMessage(response, nameof(SetStartupProfilerAsync), ValidateResponseOptions.InvalidArgumentIsRequiresSuspension);
         }
 
         /// <summary>
@@ -208,19 +191,16 @@ namespace Microsoft.Diagnostics.NETCore.Client
         /// </summary>
         public void ResumeRuntime()
         {
-            IpcMessage message = new IpcMessage(DiagnosticsServerCommandSet.Process, (byte)ProcessCommandId.ResumeRuntime);
-            var response = IpcClient.SendMessage(_endpoint, message);
-            switch ((DiagnosticsServerResponseId)response.Header.CommandId)
-            {
-                case DiagnosticsServerResponseId.Error:
-                    // Try fallback for Preview 7 and Preview 8
-                    ResumeRuntimeFallback();
-                    return;
-                case DiagnosticsServerResponseId.OK:
-                    return;
-                default:
-                    throw new ServerErrorException($"Resume runtime failed - server responded with unknown command");
-            }
+            IpcMessage request = CreateResumeRuntimeMessage();
+            IpcMessage response = IpcClient.SendMessage(_endpoint, request);
+            ValidateResponseMessage(response, nameof(ResumeRuntime));
+        }
+
+        internal async Task ResumeRuntimeAsync(CancellationToken token)
+        {
+            IpcMessage request = CreateResumeRuntimeMessage();
+            IpcMessage response = await IpcClient.SendMessageAsync(_endpoint, request, token).ConfigureAwait(false);
+            ValidateResponseMessage(response, nameof(ResumeRuntimeAsync));
         }
 
         /// <summary>
@@ -230,29 +210,16 @@ namespace Microsoft.Diagnostics.NETCore.Client
         /// <param name="value">The value of the environment variable to set.</param>
         public void SetEnvironmentVariable(string name, string value)
         {
-            if (String.IsNullOrEmpty(name))
-            {
-                throw new ArgumentException($"{nameof(name)} must be non-null.");
-            }
+            IpcMessage request = CreateSetEnvironmentVariableMessage(name, value);
+            IpcMessage response = IpcClient.SendMessage(_endpoint, request);
+            ValidateResponseMessage(response, nameof(SetEnvironmentVariable));
+        }
 
-            byte[] serializedConfiguration = SerializePayload(name, value);
-            var message = new IpcMessage(DiagnosticsServerCommandSet.Process, (byte)ProcessCommandId.SetEnvironmentVariable, serializedConfiguration);
-            var response = IpcClient.SendMessage(_endpoint, message);
-            switch ((DiagnosticsServerResponseId)response.Header.CommandId)
-            {
-                case DiagnosticsServerResponseId.Error:
-                    uint hr = BitConverter.ToUInt32(response.Payload, 0);
-                    if (hr == (uint)DiagnosticsIpcError.UnknownCommand)
-                    {
-                        throw new UnsupportedCommandException("The target runtime does not support the SetEnvironmentVariable command.");
-                    }
-
-                    throw new ServerErrorException($"SetEnvironmentVariable failed (HRESULT: 0x{hr:X8})");
-                case DiagnosticsServerResponseId.OK:
-                    return;
-                default:
-                    throw new ServerErrorException($"SetEnvironmentVariable failed - server responded with unknown command");
-            }
+        internal async Task SetEnvironmentVariableAsync(string name, string value, CancellationToken token)
+        {
+            IpcMessage request = CreateSetEnvironmentVariableMessage(name, value);
+            IpcMessage response = await IpcClient.SendMessageAsync(_endpoint, request, token).ConfigureAwait(false);
+            ValidateResponseMessage(response, nameof(SetEnvironmentVariableAsync));
         }
 
         /// <summary>
@@ -261,21 +228,22 @@ namespace Microsoft.Diagnostics.NETCore.Client
         /// <returns>A dictionary containing all of the environment variables defined in the target process.</returns>
         public Dictionary<string, string> GetProcessEnvironment()
         {
-            var message = new IpcMessage(DiagnosticsServerCommandSet.Process, (byte)ProcessCommandId.GetProcessEnvironment);
-            Stream continuation = IpcClient.SendMessage(_endpoint, message, out IpcMessage response);
-            switch ((DiagnosticsServerResponseId)response.Header.CommandId)
-            {
-                case DiagnosticsServerResponseId.Error:
-                    int hr = BitConverter.ToInt32(response.Payload, 0);
-                    throw new ServerErrorException($"Get process environment failed (HRESULT: 0x{hr:X8})");
-                case DiagnosticsServerResponseId.OK:
-                    ProcessEnvironmentHelper helper = ProcessEnvironmentHelper.Parse(response.Payload);
-                    Task<Dictionary<string, string>> envTask = helper.ReadEnvironmentAsync(continuation);
-                    envTask.Wait();
-                    return envTask.Result;
-                default:
-                    throw new ServerErrorException($"Get process environment failed - server responded with unknown command");
-            }
+            IpcMessage message = CreateProcessEnvironmentMessage();
+            using IpcResponse response = IpcClient.SendMessageGetContinuation(_endpoint, message);
+            ValidateResponseMessage(response.Message, nameof(GetProcessEnvironmentAsync));
+
+            ProcessEnvironmentHelper helper = ProcessEnvironmentHelper.Parse(response.Message.Payload);
+            return helper.ReadEnvironment(response.Continuation);
+        }
+
+        internal async Task<Dictionary<string, string>> GetProcessEnvironmentAsync(CancellationToken token)
+        {
+            IpcMessage message = CreateProcessEnvironmentMessage();
+            using IpcResponse response = await IpcClient.SendMessageGetContinuationAsync(_endpoint, message, token).ConfigureAwait(false);
+            ValidateResponseMessage(response.Message, nameof(GetProcessEnvironmentAsync));
+
+            ProcessEnvironmentHelper helper = ProcessEnvironmentHelper.Parse(response.Message.Payload);
+            return await helper.ReadEnvironmentAsync(response.Continuation, token).ConfigureAwait(false);
         }
 
         /// <summary>
@@ -304,76 +272,46 @@ namespace Microsoft.Diagnostics.NETCore.Client
             return GetAllPublishedProcesses().Distinct();
         }
 
-
-        // Fallback command for .NET 5 Preview 7 and Preview 8
-        internal void ResumeRuntimeFallback()
-        {
-            IpcMessage message = new IpcMessage(DiagnosticsServerCommandSet.Server, (byte)DiagnosticServerCommandId.ResumeRuntime);
-            var response = IpcClient.SendMessage(_endpoint, message);
-            switch ((DiagnosticsServerResponseId)response.Header.CommandId)
-            {
-                case DiagnosticsServerResponseId.Error:
-                    var hr = BitConverter.ToUInt32(response.Payload, 0);
-                    if (hr == (uint)DiagnosticsIpcError.UnknownCommand)
-                    {
-                        throw new UnsupportedCommandException($"Resume runtime command is unknown by target runtime.");
-                    }
-                    throw new ServerErrorException($"Resume runtime failed (HRESULT: 0x{hr:X8})");
-                case DiagnosticsServerResponseId.OK:
-                    return;
-                default:
-                    throw new ServerErrorException($"Resume runtime failed - server responded with unknown command");
-            }
-        }
-
         internal ProcessInfo GetProcessInfo()
         {
-            // RE: https://github.com/dotnet/runtime/issues/54083
-            // If the GetProcessInfo2 command is sent too early, it will crash the runtime instance.
-            // Disable the usage of the command until that issue is fixed.
-
             // Attempt to get ProcessInfo v2
-            //ProcessInfo processInfo = GetProcessInfo2();
-            //if (null != processInfo)
-            //{
-            //    return processInfo;
-            //}
-
-            // Attempt to get ProcessInfo v1
-            IpcMessage message = new IpcMessage(DiagnosticsServerCommandSet.Process, (byte)ProcessCommandId.GetProcessInfo);
-            var response = IpcClient.SendMessage(_endpoint, message);
-            switch ((DiagnosticsServerResponseId)response.Header.CommandId)
+            ProcessInfo processInfo = TryGetProcessInfo2();
+            if (null != processInfo)
             {
-                case DiagnosticsServerResponseId.Error:
-                    var hr = BitConverter.ToInt32(response.Payload, 0);
-                    throw new ServerErrorException($"Get process info failed (HRESULT: 0x{hr:X8})");
-                case DiagnosticsServerResponseId.OK:
-                    return ProcessInfo.ParseV1(response.Payload);
-                default:
-                    throw new ServerErrorException($"Get process info failed - server responded with unknown command");
+                return processInfo;
             }
+
+            IpcMessage request = CreateProcessInfoMessage();
+            using IpcResponse response = IpcClient.SendMessageGetContinuation(_endpoint, request);
+            return GetProcessInfoFromResponse(response, nameof(GetProcessInfo));
         }
 
-        private ProcessInfo GetProcessInfo2()
+        internal async Task<ProcessInfo> GetProcessInfoAsync(CancellationToken token)
         {
-            IpcMessage message = new IpcMessage(DiagnosticsServerCommandSet.Process, (byte)ProcessCommandId.GetProcessInfo2);
-            var response = IpcClient.SendMessage(_endpoint, message);
-            switch ((DiagnosticsServerResponseId)response.Header.CommandId)
+            // Attempt to get ProcessInfo v2
+            ProcessInfo processInfo = await TryGetProcessInfo2Async(token);
+            if (null != processInfo)
             {
-                case DiagnosticsServerResponseId.Error:
-                    uint hr = BitConverter.ToUInt32(response.Payload, 0);
-                    // In the case that the runtime doesn't understand the GetProcessInfo2 command,
-                    // just break to allow fallback to try to get ProcessInfo v1.
-                    if (hr == (uint)DiagnosticsIpcError.UnknownCommand)
-                    {
-                        return null;
-                    }
-                    throw new ServerErrorException($"GetProcessInfo2 failed (HRESULT: 0x{hr:X8})");
-                case DiagnosticsServerResponseId.OK:
-                    return ProcessInfo.ParseV2(response.Payload);
-                default:
-                    throw new ServerErrorException($"Get process info failed - server responded with unknown command");
+                return processInfo;
             }
+
+            IpcMessage request = CreateProcessInfoMessage();
+            using IpcResponse response = await IpcClient.SendMessageGetContinuationAsync(_endpoint, request, token).ConfigureAwait(false);
+            return GetProcessInfoFromResponse(response, nameof(GetProcessInfoAsync));
+        }
+
+        private ProcessInfo TryGetProcessInfo2()
+        {
+            IpcMessage request = CreateProcessInfo2Message();
+            using IpcResponse response2 = IpcClient.SendMessageGetContinuation(_endpoint, request);
+            return TryGetProcessInfo2FromResponse(response2, nameof(GetProcessInfo));
+        }
+
+        private async Task<ProcessInfo> TryGetProcessInfo2Async(CancellationToken token)
+        {
+            IpcMessage request = CreateProcessInfo2Message();
+            using IpcResponse response2 = await IpcClient.SendMessageGetContinuationAsync(_endpoint, request, token).ConfigureAwait(false);
+            return TryGetProcessInfo2FromResponse(response2, nameof(GetProcessInfoAsync));
         }
 
         private static byte[] SerializePayload<T>(T arg)
@@ -450,10 +388,143 @@ namespace Microsoft.Diagnostics.NETCore.Client
                 uint uiValue = bValue ? (uint)1 : 0;
                 writer.Write(uiValue);
             }
+            else if (typeof(T) == typeof(Guid))
+            {
+                Guid guidVal = (Guid)((object)obj);
+                writer.Write(guidVal.ToByteArray());
+            }
             else
             {
                 throw new ArgumentException($"Type {obj.GetType()} is not supported in SerializePayloadArgument, please add it.");
             }
+        }
+
+        private static IpcMessage CreateAttachProfilerMessage(TimeSpan attachTimeout, Guid profilerGuid, string profilerPath, byte[] additionalData)
+        {
+            if (profilerGuid == null || profilerGuid == Guid.Empty)
+            {
+                throw new ArgumentException($"{nameof(profilerGuid)} must be a valid Guid");
+            }
+
+            if (String.IsNullOrEmpty(profilerPath))
+            {
+                throw new ArgumentException($"{nameof(profilerPath)} must be non-null");
+            }
+
+            byte[] serializedConfiguration = SerializePayload((uint)attachTimeout.TotalSeconds, profilerGuid, profilerPath, additionalData);
+            return new IpcMessage(DiagnosticsServerCommandSet.Profiler, (byte)ProfilerCommandId.AttachProfiler, serializedConfiguration);
+        }
+
+        private static IpcMessage CreateProcessEnvironmentMessage()
+        {
+            return new IpcMessage(DiagnosticsServerCommandSet.Process, (byte)ProcessCommandId.GetProcessEnvironment);
+        }
+
+        private static IpcMessage CreateProcessInfoMessage()
+        {
+            return new IpcMessage(DiagnosticsServerCommandSet.Process, (byte)ProcessCommandId.GetProcessInfo);
+        }
+
+        private static IpcMessage CreateProcessInfo2Message()
+        {
+            return new IpcMessage(DiagnosticsServerCommandSet.Process, (byte)ProcessCommandId.GetProcessInfo2);
+        }
+
+        private static IpcMessage CreateResumeRuntimeMessage()
+        {
+            return new IpcMessage(DiagnosticsServerCommandSet.Process, (byte)ProcessCommandId.ResumeRuntime);
+        }
+
+        private static IpcMessage CreateSetEnvironmentVariableMessage(string name, string value)
+        {
+            if (String.IsNullOrEmpty(name))
+            {
+                throw new ArgumentException($"{nameof(name)} must be non-null.");
+            }
+
+            byte[] serializedConfiguration = SerializePayload(name, value);
+            return new IpcMessage(DiagnosticsServerCommandSet.Process, (byte)ProcessCommandId.SetEnvironmentVariable, serializedConfiguration);
+        }
+
+        private static IpcMessage CreateSetStartupProfilerMessage(Guid profilerGuid, string profilerPath)
+        {
+            if (profilerGuid == null || profilerGuid == Guid.Empty)
+            {
+                throw new ArgumentException($"{nameof(profilerGuid)} must be a valid Guid");
+            }
+
+            if (String.IsNullOrEmpty(profilerPath))
+            {
+                throw new ArgumentException($"{nameof(profilerPath)} must be non-null");
+            }
+
+            byte[] serializedConfiguration = SerializePayload(profilerGuid, profilerPath);
+            return new IpcMessage(DiagnosticsServerCommandSet.Profiler, (byte)ProfilerCommandId.StartupProfiler, serializedConfiguration);
+        }
+
+        private static IpcMessage CreateWriteDumpMessage(DumpType dumpType, string dumpPath, bool logDumpGeneration)
+        {
+            if (string.IsNullOrEmpty(dumpPath))
+                throw new ArgumentNullException($"{nameof(dumpPath)} required");
+
+            byte[] payload = SerializePayload(dumpPath, (uint)dumpType, logDumpGeneration);
+            return new IpcMessage(DiagnosticsServerCommandSet.Dump, (byte)DumpCommandId.GenerateCoreDump, payload);
+        }
+
+        private static ProcessInfo GetProcessInfoFromResponse(IpcResponse response, string operationName)
+        {
+            ValidateResponseMessage(response.Message, operationName);
+
+            return ProcessInfo.ParseV1(response.Message.Payload);
+        }
+
+        private static ProcessInfo TryGetProcessInfo2FromResponse(IpcResponse response, string operationName)
+        {
+            if (!ValidateResponseMessage(response.Message, operationName, ValidateResponseOptions.UnknownCommandReturnsFalse))
+            {
+                return null;
+            }
+
+            return ProcessInfo.ParseV2(response.Message.Payload);
+        }
+
+        internal static bool ValidateResponseMessage(IpcMessage responseMessage, string operationName, ValidateResponseOptions options = ValidateResponseOptions.None)
+        {
+            switch ((DiagnosticsServerResponseId)responseMessage.Header.CommandId)
+            {
+                case DiagnosticsServerResponseId.Error:
+                    uint hr = BitConverter.ToUInt32(responseMessage.Payload, 0);
+                    switch (hr)
+                    {
+                        case (uint)DiagnosticsIpcError.UnknownCommand:
+                            if (options.HasFlag(ValidateResponseOptions.UnknownCommandReturnsFalse))
+                            {
+                                return false;
+                            }
+                            throw new UnsupportedCommandException($"{operationName} failed - Command is not supported.");
+                        case (uint)DiagnosticsIpcError.ProfilerAlreadyActive:
+                            throw new ProfilerAlreadyActiveException($"{operationName} failed - A profiler is already loaded.");
+                        case (uint)DiagnosticsIpcError.InvalidArgument:
+                            if (options.HasFlag(ValidateResponseOptions.InvalidArgumentIsRequiresSuspension))
+                            {
+                                throw new ServerErrorException($"{operationName} failed - The runtime must be suspended for this command.");
+                            }
+                            throw new UnsupportedCommandException($"{operationName} failed - Invalid command argument.");
+                    }
+                    throw new ServerErrorException($"{operationName} failed - HRESULT: 0x{hr:X8}");
+                case DiagnosticsServerResponseId.OK:
+                    return true;
+                default:
+                    throw new ServerErrorException($"{operationName} failed - Server responded with unknown response.");
+            }
+        }
+
+        [Flags]
+        internal enum ValidateResponseOptions
+        {
+            None = 0x0,
+            UnknownCommandReturnsFalse = 0x1,
+            InvalidArgumentIsRequiresSuspension = 0x2,
         }
     }
 }
