@@ -6,55 +6,114 @@ using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
+using System.Threading;
+using System.Threading.Tasks;
 
 namespace Microsoft.Diagnostics.NETCore.Client
 {
     public class EventPipeSession : IDisposable
     {
-        private IEnumerable<EventPipeProvider> _providers;
-        private bool _requestRundown;
-        private int _circularBufferMB;
         private long _sessionId;
         private IpcEndpoint _endpoint;
         private bool _disposedValue = false; // To detect redundant calls
         private bool _stopped = false; // To detect redundant calls
+        private readonly IpcResponse _response;
 
-        internal EventPipeSession(IpcEndpoint endpoint, IEnumerable<EventPipeProvider> providers, bool requestRundown, int circularBufferMB)
+        private EventPipeSession(IpcEndpoint endpoint, IpcResponse response, long sessionId)
         {
             _endpoint = endpoint;
-            _providers = providers;
-            _requestRundown = requestRundown;
-            _circularBufferMB = circularBufferMB;
-            
-            var config = new EventPipeSessionConfiguration(circularBufferMB, EventPipeSerializationFormat.NetTrace, providers, requestRundown);
-            var message = new IpcMessage(DiagnosticsServerCommandSet.EventPipe, (byte)EventPipeCommandId.CollectTracing2, config.SerializeV2());
-            EventStream = IpcClient.SendMessage(endpoint, message, out var response);
-            switch ((DiagnosticsServerResponseId)response.Header.CommandId)
-            {
-                case DiagnosticsServerResponseId.OK:
-                    _sessionId = BitConverter.ToInt64(response.Payload, 0);
-                    break;
-                case DiagnosticsServerResponseId.Error:
-                    var hr = BitConverter.ToInt32(response.Payload, 0);
-                    throw new ServerErrorException($"EventPipe session start failed (HRESULT: 0x{hr:X8})");
-                default:
-                    throw new ServerErrorException($"EventPipe session start failed - Server responded with unknown command");
-            }
+            _response = response;
+            _sessionId = sessionId;
         }
 
-        public Stream EventStream { get; }
+        public Stream EventStream => _response.Continuation;
+
+        internal static EventPipeSession Start(IpcEndpoint endpoint, IEnumerable<EventPipeProvider> providers, bool requestRundown, int circularBufferMB)
+        {
+            IpcMessage requestMessage = CreateStartMessage(providers, requestRundown, circularBufferMB);
+            IpcResponse? response = IpcClient.SendMessageGetContinuation(endpoint, requestMessage);
+            return CreateSessionFromResponse(endpoint, ref response, nameof(Start));
+        }
+
+        internal static async Task<EventPipeSession> StartAsync(IpcEndpoint endpoint, IEnumerable<EventPipeProvider> providers, bool requestRundown, int circularBufferMB, CancellationToken cancellationToken)
+        {
+            IpcMessage requestMessage = CreateStartMessage(providers, requestRundown, circularBufferMB);
+            IpcResponse? response = await IpcClient.SendMessageGetContinuationAsync(endpoint, requestMessage, cancellationToken).ConfigureAwait(false);
+            return CreateSessionFromResponse(endpoint, ref response, nameof(StartAsync));
+        }
 
         ///<summary>
         /// Stops the given session
         ///</summary>
         public void Stop()
         {
+            if (TryCreateStopMessage(out IpcMessage requestMessage))
+            {
+                try
+                {
+                    IpcMessage response = IpcClient.SendMessage(_endpoint, requestMessage);
+
+                    DiagnosticsClient.ValidateResponseMessage(response, nameof(Stop));
+                }
+                // On non-abrupt exits (i.e. the target process has already exited and pipe is gone, sending Stop command will fail).
+                catch (IOException)
+                {
+                    throw new ServerNotAvailableException("Could not send Stop command. The target process may have exited.");
+                }
+            }
+        }
+
+        public async Task StopAsync(CancellationToken cancellationToken)
+        {
+            if (TryCreateStopMessage(out IpcMessage requestMessage))
+            {
+                try
+                {
+                    IpcMessage response = await IpcClient.SendMessageAsync(_endpoint, requestMessage, cancellationToken).ConfigureAwait(false);
+
+                    DiagnosticsClient.ValidateResponseMessage(response, nameof(StopAsync));
+                }
+                // On non-abrupt exits (i.e. the target process has already exited and pipe is gone, sending Stop command will fail).
+                catch (IOException)
+                {
+                    throw new ServerNotAvailableException("Could not send Stop command. The target process may have exited.");
+                }
+            }
+        }
+
+        private static IpcMessage CreateStartMessage(IEnumerable<EventPipeProvider> providers, bool requestRundown, int circularBufferMB)
+        {
+            var config = new EventPipeSessionConfiguration(circularBufferMB, EventPipeSerializationFormat.NetTrace, providers, requestRundown);
+            return new IpcMessage(DiagnosticsServerCommandSet.EventPipe, (byte)EventPipeCommandId.CollectTracing2, config.SerializeV2());
+        }
+
+        private static EventPipeSession CreateSessionFromResponse(IpcEndpoint endpoint, ref IpcResponse? response, string operationName)
+        {
+            try
+            {
+                DiagnosticsClient.ValidateResponseMessage(response.Value.Message, operationName);
+
+                long sessionId = BitConverter.ToInt64(response.Value.Message.Payload, 0);
+
+                var session = new EventPipeSession(endpoint, response.Value, sessionId);
+                response = null;
+                return session;
+            }
+            finally
+            {
+                response?.Dispose();
+            }
+        }
+
+        private bool TryCreateStopMessage(out IpcMessage stopMessage)
+        {
             Debug.Assert(_sessionId > 0);
-            
+
             // Do not issue another Stop command if it has already been issued for this session instance.
             if (_stopped)
             {
-                return;
+                stopMessage = null;
+                return false;
             }
             else
             {
@@ -62,27 +121,10 @@ namespace Microsoft.Diagnostics.NETCore.Client
             }
 
             byte[] payload = BitConverter.GetBytes(_sessionId);
-            IpcMessage response;
-            try
-            {
-                response = IpcClient.SendMessage(_endpoint, new IpcMessage(DiagnosticsServerCommandSet.EventPipe, (byte)EventPipeCommandId.StopTracing, payload));
-            }
-            // On non-abrupt exits (i.e. the target process has already exited and pipe is gone, sending Stop command will fail).
-            catch (IOException)
-            {
-                throw new ServerNotAvailableException("Could not send Stop command. The target process may have exited.");
-            }
 
-            switch ((DiagnosticsServerResponseId)response.Header.CommandId)
-            {
-                case DiagnosticsServerResponseId.OK:
-                    return;
-                case DiagnosticsServerResponseId.Error:
-                    var hr = BitConverter.ToInt32(response.Payload, 0);
-                    throw new ServerErrorException($"EventPipe session stop failed (HRESULT: 0x{hr:X8})");
-                default:
-                    throw new ServerErrorException($"EventPipe session stop failed - Server responded with unknown command");
-            }
+            stopMessage = new IpcMessage(DiagnosticsServerCommandSet.EventPipe, (byte)EventPipeCommandId.StopTracing, payload);
+
+            return true;
         }
 
         protected virtual void Dispose(bool disposing)
@@ -101,7 +143,7 @@ namespace Microsoft.Diagnostics.NETCore.Client
             {
                 if (disposing)
                 {
-                    EventStream?.Dispose();
+                    _response.Dispose();
                 }
                 _disposedValue = true;
             }
