@@ -51,7 +51,7 @@ namespace Microsoft.Diagnostics.TestHelpers
 
         protected override string GetFramework(TestConfiguration config)
         {
-            return config.BuildProjectFramework ?? "netcoreapp2.1";
+            return config.BuildProjectFramework ?? "netcoreapp3.1";
         }
 
         protected override string GetDebuggeeBinaryDirPath(string debuggeeProjectDirPath, string framework, string runtime)
