@@ -8677,7 +8677,6 @@ DECLARE_API(ThreadPool)
     do // while (false)
     {
         UINT64 ui64Value = 0;
-        UINT32 ui32Value = 0;
 
         // Determine if the portable thread pool is enabled
         if (FAILED(
@@ -8774,12 +8773,12 @@ DECLARE_API(ThreadPool)
             accumulatedOffset += offset;
 
             offset = GetValueFieldOffset(vCountsField.MTOfType, W("_data"));
-            if (offset < 0 || FAILED(MOVE(ui32Value, cdaTpInstance + accumulatedOffset + offset)))
+            if (offset < 0 || FAILED(MOVE(ui64Value, cdaTpInstance + accumulatedOffset + offset)))
             {
                 ExtOut("    %s\n", "Failed to read PortableThreadPool._separated.counts._data");
                 break;
             }
-            UINT32 data = ui32Value;
+            UINT64 data = ui64Value;
 
             const UINT8 NumProcessingWorkShift = 0;
             const UINT8 NumExistingThreadsShift = 16;
@@ -9107,6 +9106,7 @@ DECLARE_API(ThreadPool)
                 "Stabilizing",
                 "Starvation",
                 "ThreadTimedOut",
+                "CooperativeBlocking",
                 "Undefined"
             };
 
@@ -16749,6 +16749,10 @@ DECLARE_API(SetSymbolServer)
     else if (loadNative)
     {
         Status = LoadNativeSymbols();
+        if (FAILED(Status))
+        {
+            ExtErr("Symbol server not set\n");
+        }
     }
 #endif
     else
