@@ -8,102 +8,60 @@ using System.Text;
 
 namespace Microsoft.Diagnostics.Tools.Counters
 {
-    public interface ICounterPayload
+    public class CounterPayload
     {
-        string GetName();
-        double GetValue();
-        string GetDisplay();
-        string GetCounterType();
+        public CounterPayload(string providerName, string name, string displayName, string displayUnits, string tags, double value, DateTime timestamp, string type)
+        {
+            ProviderName = providerName;
+            Name = name;
+            Tags = tags;
+            Value = value;
+            Timestamp = timestamp;
+            CounterType = type;
+        }
+
+        public string ProviderName { get; private set; }
+        public string Name { get; private set; }
+        public double Value { get; private set; }
+        public virtual string DisplayName { get; protected set; }
+        public string CounterType { get; private set; }
+        public DateTime Timestamp { get; private set; }
+        public string Tags { get; private set; }
     }
 
 
-    class CounterPayload : ICounterPayload
+    class GaugePayload : CounterPayload
     {
-        public string m_Name;
-        public double m_Value;
-        public string m_DisplayName;
-        public string m_DisplayUnits;
-
-        public CounterPayload(IDictionary<string, object> payloadFields)
+        public GaugePayload(string providerName, string name, string displayName, string displayUnits, string tags, double value, DateTime timestamp) :
+            base(providerName, name, displayName, displayUnits, tags, value, timestamp, "Metric")
         {
-            m_Name = payloadFields["Name"].ToString();
-            m_Value = (double)payloadFields["Mean"];
-            m_DisplayName = payloadFields["DisplayName"].ToString();
-            m_DisplayUnits = payloadFields["DisplayUnits"].ToString();
-
             // In case these properties are not provided, set them to appropriate values.
-            m_DisplayName = m_DisplayName.Length == 0 ? m_Name : m_DisplayName;
-        }
-
-        public string GetName()
-        {
-            return m_Name;
-        }
-
-        public double GetValue()
-        {
-            return m_Value;
-        }
-
-        public string GetDisplay()
-        {
-            if (m_DisplayUnits.Length > 0)
-            {
-                return $"{m_DisplayName} ({m_DisplayUnits})";
-            }
-            return $"{m_DisplayName}";
-        }
-
-        public string GetCounterType()
-        {
-            return "Metric";
+            string counterName = string.IsNullOrEmpty(displayName) ? name : displayName;
+            DisplayName = !string.IsNullOrEmpty(displayUnits) ? $"{counterName} ({displayUnits})" : counterName;
         }
     }
 
-    class IncrementingCounterPayload : ICounterPayload
+    class RatePayload : CounterPayload
     {
-        public string m_Name;
-        public double m_Value;
-        public string m_DisplayName;
-        public string m_Interval;
-        public string m_DisplayUnits;
-
-        public IncrementingCounterPayload(IDictionary<string, object> payloadFields, int interval)
+        public RatePayload(string providerName, string name, string displayName, string displayUnits, string tags, double value, double intervalSecs, DateTime timestamp) :
+            base(providerName, name, displayName, displayUnits, tags, value, timestamp, "Rate")
         {
-            m_Name = payloadFields["Name"].ToString();
-            m_Value = (double)payloadFields["Increment"];
-            m_DisplayName = payloadFields["DisplayName"].ToString();
-            m_DisplayUnits = payloadFields["DisplayUnits"].ToString();
-            m_Interval = interval.ToString() + " sec";
-
             // In case these properties are not provided, set them to appropriate values.
-            m_DisplayName = m_DisplayName.Length == 0 ? m_Name : m_DisplayName;
+            string counterName = string.IsNullOrEmpty(displayName) ? name : displayName;
+            string unitsName = string.IsNullOrEmpty(displayUnits) ? "Count" : displayUnits;
+            string intervalName = intervalSecs.ToString() + " sec";
+            DisplayName = $"{counterName} ({unitsName} / {intervalName})";
         }
+    }
 
-        public string GetName()
+    class PercentilePayload : CounterPayload
+    {
+        public PercentilePayload(string providerName, string name, string displayName, string displayUnits, string tags, double val, DateTime timestamp) :
+            base(providerName, name, displayName, displayUnits, tags, val, timestamp, "Metric")
         {
-            return m_Name;
-        }
-
-        public double GetValue()
-        {
-            return m_Value;
-        }
-
-        public string GetDisplay()
-        {
-            return $"{m_DisplayName} ({GetDisplayUnits()} / {m_Interval})";
-        }
-
-        public string GetCounterType()
-        {
-            return "Rate";
-        }
-
-        private string GetDisplayUnits()
-        {
-            if (m_DisplayUnits.Length == 0) return "Count";
-            return m_DisplayUnits;
+            // In case these properties are not provided, set them to appropriate values.
+            string counterName = string.IsNullOrEmpty(displayName) ? name : displayName;
+            DisplayName = !string.IsNullOrEmpty(displayUnits) ? $"{counterName} ({displayUnits})" : counterName;
         }
     }
 }
