@@ -6,6 +6,7 @@ using System;
 using System.Collections.Generic;
 using System.Globalization;
 using System.Linq;
+using System.Text;
 
 namespace Microsoft.Diagnostics.Tools.Counters.Exporters
 {
@@ -35,17 +36,34 @@ namespace Microsoft.Diagnostics.Tools.Counters.Exporters
             public ObservedCounter(string displayName) => DisplayName = displayName;
             public string DisplayName { get; } // Display name for this counter.
             public int Row { get; set; } // Assigned row for this counter. May change during operation.
+            public Dictionary<string, ObservedTagSet> TagSets { get; } = new Dictionary<string, ObservedTagSet>();
+
+            public bool RenderValueInline => TagSets.Count == 0 ||
+                       (TagSets.Count == 1 && string.IsNullOrEmpty(TagSets.Keys.First()));
+            public double LastValue { get; set; }
+        }
+
+        private class ObservedTagSet
+        {
+            public ObservedTagSet(string tags)
+            {
+                Tags = tags;
+            }
+            public string Tags { get; }
+            public string DisplayTags => string.IsNullOrEmpty(Tags) ? "<no tags>" : Tags;
+            public int Row { get; set; } // Assigned row for this counter. May change during operation.
+            public double LastValue { get; set; }
         }
 
         private readonly object _lock = new object();
         private readonly Dictionary<string, ObservedProvider> providers = new Dictionary<string, ObservedProvider>(); // Tracks observed providers and counters.
         private const int Indent = 4; // Counter name indent size.
         private int maxNameLength = 40; // Allow room for 40 character counter names by default.
-        private int maxPreDecimalDigits = 11; // Allow room for values up to 999 million by default.
 
         private int STATUS_ROW; // Row # of where we print the status of dotnet-counters
         private bool paused = false;
         private bool initialized = false;
+        private string _errorText = null;
 
         private int maxRow = -1;
 
@@ -57,6 +75,12 @@ namespace Microsoft.Diagnostics.Tools.Counters.Exporters
         public void EventPipeSourceConnected()
         {
             // Do nothing
+        }
+
+        public void SetErrorText(string errorText)
+        {
+            _errorText = errorText;
+            AssignRowsAndInitializeDisplay();
         }
 
         private void UpdateStatus()
@@ -74,6 +98,11 @@ namespace Microsoft.Diagnostics.Tools.Counters.Exporters
             int row = Console.CursorTop;
             Console.WriteLine("Press p to pause, r to resume, q to quit."); row++;
             Console.WriteLine($"    Status: {GetStatus()}");                STATUS_ROW = row++;
+            if (_errorText != null)
+            {
+                Console.WriteLine(_errorText);
+                row += GetLineWrappedLines(_errorText);
+            }
             Console.WriteLine();                                            row++; // Blank line.
 
             foreach (ObservedProvider provider in providers.Values.OrderBy(p => p.KnownProvider == null).ThenBy(p => p.Name)) // Known providers first.
@@ -81,8 +110,22 @@ namespace Microsoft.Diagnostics.Tools.Counters.Exporters
                 Console.WriteLine($"[{provider.Name}]"); row++;
                 foreach (ObservedCounter counter in provider.Counters.Values.OrderBy(c => c.DisplayName))
                 {
-                    Console.WriteLine($"{new string(' ', Indent)}{counter.DisplayName}");
+                    string name = MakeFixedWidth($"{new string(' ', Indent)}{counter.DisplayName}", Indent + maxNameLength);
                     counter.Row = row++;
+                    if (counter.RenderValueInline)
+                    {
+                        Console.WriteLine($"{name} {FormatValue(counter.LastValue)}");
+                    }
+                    else
+                    {
+                        Console.WriteLine(name);
+                        foreach (ObservedTagSet tagSet in counter.TagSets.Values.OrderBy(t => t.Tags))
+                        {
+                            string tagName = MakeFixedWidth($"{new string(' ', 2 * Indent)}{tagSet.Tags}", Indent + maxNameLength);
+                            Console.WriteLine($"{tagName} {FormatValue(tagSet.LastValue)}");
+                            tagSet.Row = row++;
+                        }
+                    }
                 }
             }
 
@@ -100,7 +143,7 @@ namespace Microsoft.Diagnostics.Tools.Counters.Exporters
             UpdateStatus();
         }
 
-        public void CounterPayloadReceived(string providerName, ICounterPayload payload, bool pauseCmdSet)
+        public void CounterPayloadReceived(CounterPayload payload, bool pauseCmdSet)
         {
             lock (_lock)
             {
@@ -115,7 +158,9 @@ namespace Microsoft.Diagnostics.Tools.Counters.Exporters
                     return;
                 }
 
-                string name = payload.GetName();
+                string providerName = payload.ProviderName;
+                string name = payload.Name;
+                string tags = payload.Tags;
 
                 bool redraw = false;
                 if (!providers.TryGetValue(providerName, out ObservedProvider provider))
@@ -126,23 +171,22 @@ namespace Microsoft.Diagnostics.Tools.Counters.Exporters
 
                 if (!provider.Counters.TryGetValue(name, out ObservedCounter counter))
                 {
-                    string displayName = payload.GetDisplay();
+                    string displayName = payload.DisplayName;
                     provider.Counters[name] = counter = new ObservedCounter(displayName);
                     maxNameLength = Math.Max(maxNameLength, displayName.Length);
+                    if(tags != null)
+                    {
+                        counter.LastValue = payload.Value;
+                    }
                     redraw = true;
                 }
 
-                const string DecimalPlaces = "###";
-                string payloadVal = payload.GetValue().ToString("#,0." + DecimalPlaces, CultureInfo.CurrentCulture);
-                int decimalIndex = payloadVal.IndexOf(CultureInfo.CurrentCulture.NumberFormat.NumberDecimalSeparator, StringComparison.CurrentCulture);
-                if (decimalIndex == -1)
+                ObservedTagSet tagSet = null;
+                if (tags != null && !counter.TagSets.TryGetValue(tags, out tagSet))
                 {
-                    decimalIndex = payloadVal.Length;
-                }
-
-                if (decimalIndex > maxPreDecimalDigits)
-                {
-                    maxPreDecimalDigits = decimalIndex;
+                    counter.TagSets[tags] = tagSet = new ObservedTagSet(tags);
+                    maxNameLength = Math.Max(maxNameLength, tagSet.DisplayTags.Length);
+                    tagSet.LastValue = payload.Value;
                     redraw = true;
                 }
 
@@ -151,10 +195,77 @@ namespace Microsoft.Diagnostics.Tools.Counters.Exporters
                     AssignRowsAndInitializeDisplay();
                 }
 
-                Console.SetCursorPosition(Indent + maxNameLength + 1, counter.Row);
-                int prefixSpaces = maxPreDecimalDigits - decimalIndex;
-                int postfixSpaces = DecimalPlaces.Length - (payloadVal.Length - decimalIndex - 1);
-                Console.Write($"{new string(' ', prefixSpaces)}{payloadVal}{new string(' ', postfixSpaces)}");
+                int row = counter.RenderValueInline ? counter.Row : tagSet.Row;
+                Console.SetCursorPosition(Indent + maxNameLength + 1, row);
+                Console.Write(FormatValue(payload.Value));
+            }
+        }
+
+        private static int GetLineWrappedLines(string text)
+        {
+            string[] lines = text.Split(Environment.NewLine);
+            int lineCount = lines.Length;
+            int width = Console.BufferWidth;
+            foreach(string line in lines)
+            {
+                lineCount += (int)Math.Floor(((float)line.Length) / width);
+            }
+            return lineCount;
+        }
+
+        private string FormatValue(double value)
+        {
+            string valueText = null;
+            // The value field is one of:
+            //  a) If abs(value) >= 10^9 then it is formatted as 0.####e+00
+            //  b) otherwise leading - or space, 10 leading digits with separators (or spaces), decimal separator or space,
+            //     3 decimal digits or spaces.
+            //
+            // For example:              
+            //   1,421,893.21    
+            //           0.123
+            //          -0.123
+            //  4.9012e+25
+            //    -675,430.9    
+            // -12,675,430.9  
+            //           7
+
+
+            if (Math.Abs(value) >= 100_000_000)
+            {
+                valueText = string.Format(CultureInfo.CurrentCulture, "{0,15:0.####e+00}   ", value);
+            }
+            else
+            {
+                string formattedVal = value.ToString("##,###,##0.###");
+                int seperatorIndex = formattedVal.IndexOf(CultureInfo.CurrentCulture.NumberFormat.NumberDecimalSeparator);
+                if (seperatorIndex == -1)
+                {
+                    formattedVal += "    ";
+                }
+                else
+                {
+                    int decimalDigits = formattedVal.Length - 1 - seperatorIndex;
+                    formattedVal += new string(' ', 3 - decimalDigits);
+                }
+                valueText = string.Format(CultureInfo.CurrentCulture, "{0,15}", formattedVal);
+            }
+            return valueText;
+        }
+
+        private static string MakeFixedWidth(string text, int width)
+        {
+            if(text.Length == width)
+            {
+                return text;
+            }
+            else if(text.Length > width)
+            {
+                return text.Substring(0, width);
+            }
+            else
+            {
+                return text += new string(' ', width - text.Length);
             }
         }
 
