@@ -51,12 +51,18 @@ namespace Microsoft.Diagnostics.Tools.Counters.Exporters
         {
             Console.WriteLine("Starting a counter session. Press Q to quit.");
         }
+
+        public void SetErrorText(string errorText)
+        {
+            Console.WriteLine(errorText);
+        }
+
         public void ToggleStatus(bool paused)
         {
             // Do nothing
         }
 
-        public void CounterPayloadReceived(string providerName, ICounterPayload payload, bool _)
+        public void CounterPayloadReceived(CounterPayload payload, bool _)
         {
             lock (_lock)
             {
@@ -65,13 +71,13 @@ namespace Microsoft.Diagnostics.Tools.Counters.Exporters
                     File.AppendAllText(_output, builder.ToString());
                     builder.Clear();
                 }
-
                 builder
                     .Append("{ \"timestamp\": \"").Append(DateTime.Now.ToString("u")).Append("\", ")
-                    .Append(" \"provider\": \"").Append(providerName).Append("\", ")
-                    .Append(" \"name\": \"").Append(payload.GetDisplay()).Append("\", ")
-                    .Append(" \"counterType\": \"").Append(payload.GetCounterType()).Append("\", ")
-                    .Append(" \"value\": ").Append(payload.GetValue().ToString(CultureInfo.InvariantCulture)).Append(" },");
+                    .Append(" \"provider\": \"").Append(JsonEscape(payload.ProviderName)).Append("\", ")
+                    .Append(" \"name\": \"").Append(JsonEscape(payload.DisplayName)).Append("\", ")
+                    .Append(" \"tags\": \"").Append(JsonEscape(payload.Tags)).Append("\", ")
+                    .Append(" \"counterType\": \"").Append(JsonEscape(payload.CounterType)).Append("\", ")
+                    .Append(" \"value\": ").Append(payload.Value.ToString(CultureInfo.InvariantCulture)).Append(" },");
             }
         }
 
@@ -85,6 +91,55 @@ namespace Microsoft.Diagnostics.Tools.Counters.Exporters
                 File.AppendAllText(_output, builder.ToString());
             }
             Console.WriteLine("File saved to " + _output);
+        }
+
+        static readonly char[] s_escapeChars = new char[] { '"', '\n', '\r', '\t', '\\', '\b', '\f' };
+
+        private string JsonEscape(string input)
+        {
+            int offset = input.IndexOfAny(s_escapeChars);
+            if(offset == -1)
+            {
+                // fast path
+                return input;
+            }
+
+            // slow path
+            // this could be written more efficiently but I expect it to be quite rare and not performance sensitive
+            // so I didn't feel justified writing a complex routine or adding a few 100KB for a dependency on a
+            // better performing JSON library
+            StringBuilder sb = new StringBuilder(input.Length + 10);
+            foreach(char c in input)
+            {
+                switch (c)
+                {
+                    case '\"':
+                        sb.Append("\\\"");
+                        break;
+                    case '\n':
+                        sb.Append("\\n");
+                        break;
+                    case '\r':
+                        sb.Append("\\r");
+                        break;
+                    case '\t':
+                        sb.Append("\\t");
+                        break;
+                    case '\\':
+                        sb.Append("\\\\");
+                        break;
+                    case '\b':
+                        sb.Append("\\b");
+                        break;
+                    case '\f':
+                        sb.Append("\\f");
+                        break;
+                    default:
+                        sb.Append(c);
+                        break;
+                }
+            }
+            return sb.ToString();
         }
     }
 }
