@@ -208,6 +208,9 @@ namespace Microsoft.Diagnostics.Tools.DiagnosticsServerRouter
                 }
             }
 
+            if (string.IsNullOrEmpty(ipcServer))
+                ipcServer = GetDefaultIpcServerPath(logger);
+
             var routerTask = DiagnosticsServerRouterRunner.runIpcServerTcpClientRouter(linkedCancelToken.Token, ipcServer, tcpClient, runtimeTimeout == Timeout.Infinite ? runtimeTimeout : runtimeTimeout * 1000, tcpClientRouterFactory, logger, Launcher);
 
             while (!linkedCancelToken.IsCancellationRequested)
@@ -338,7 +341,7 @@ namespace Microsoft.Diagnostics.Tools.DiagnosticsServerRouter
                 message.Append("WARNING: Binding tcp server endpoint to anything except loopback interface ");
                 message.Append("(localhost, 127.0.0.1 or [::1]) is NOT recommended. Any connections towards ");
                 message.Append("tcp server endpoint will be unauthenticated and unencrypted. This component ");
-                message.Append("is intented for development use and should only be run in development and ");
+                message.Append("is intended for development use and should only be run in development and ");
                 message.Append("testing environments.");
                 message.AppendLine();
 
