@@ -2,21 +2,13 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using Microsoft.Diagnostics.Tracing;
 using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.Diagnostics.Tracing;
-using System.IO;
-using System.Runtime.InteropServices;
-using System.Text.RegularExpressions;
-using System.Threading;
 using System.Threading.Tasks;
 using Xunit;
 using Xunit.Abstractions;
-
-using Microsoft.Diagnostics.Tracing;
-using Microsoft.Diagnostics.TestHelpers;
-using Microsoft.Diagnostics.NETCore.Client;
 
 namespace Microsoft.Diagnostics.NETCore.Client
 {
@@ -29,16 +21,28 @@ namespace Microsoft.Diagnostics.NETCore.Client
             output = outputHelper;
         }
 
+        [Fact]
+        public Task BasicEventPipeSessionTest()
+        {
+            return BasicEventPipeSessionTestCore(useAsync: false);
+        }
+
+        [Fact]
+        public Task BasicEventPipeSessionTestAsync()
+        {
+            return BasicEventPipeSessionTestCore(useAsync: true);
+        }
+
         /// <summary>
         /// A simple test that checks if we can create an EventPipeSession on a child process
         /// </summary>
-        [Fact]
-        public void BasicEventPipeSessionTest()
+        private async Task BasicEventPipeSessionTestCore(bool useAsync)
+
         {
             using TestRunner runner = new TestRunner(CommonHelper.GetTraceePathWithArgs(), output);
             runner.Start(timeoutInMSPipeCreation: 15_000, testProcessTimeout: 60_000);
-            DiagnosticsClient client = new DiagnosticsClient(runner.Pid);
-            using (var session = client.StartEventPipeSession(new List<EventPipeProvider>()
+            DiagnosticsClientApiShim clientShim = new DiagnosticsClientApiShim(new DiagnosticsClient(runner.Pid), useAsync);
+            using (var session = await clientShim.StartEventPipeSession(new List<EventPipeProvider>()
             {
                 new EventPipeProvider("Microsoft-Windows-DotNETRuntime", EventLevel.Informational)
             }))
@@ -48,18 +52,29 @@ namespace Microsoft.Diagnostics.NETCore.Client
             runner.Stop();
         }
 
+        [Fact]
+        public Task EventPipeSessionStreamTest()
+        {
+            return EventPipeSessionStreamTestCore(useAsync: false);
+        }
+
+        [Fact]
+        public Task EventPipeSessionStreamTestAsync()
+        {
+            return EventPipeSessionStreamTestCore(useAsync: true);
+        }
+
         /// <summary>
         /// Checks if we can create an EventPipeSession and can get some expected events out of it.
         /// </summary>
-        [Fact]
-        public void EventPipeSessionStreamTest()
+        private async Task EventPipeSessionStreamTestCore(bool useAsync)
         {
             TestRunner runner = new TestRunner(CommonHelper.GetTraceePathWithArgs(), output);
             runner.Start(timeoutInMSPipeCreation: 15_000, testProcessTimeout: 60_000);
-            DiagnosticsClient client = new DiagnosticsClient(runner.Pid);
+            DiagnosticsClientApiShim clientShim = new DiagnosticsClientApiShim(new DiagnosticsClient(runner.Pid), useAsync);
             runner.PrintStatus();
             output.WriteLine($"[{DateTime.Now.ToString()}] Trying to start an EventPipe session on process {runner.Pid}");
-            using (var session = client.StartEventPipeSession(new List<EventPipeProvider>()
+            using (var session = await clientShim.StartEventPipeSession(new List<EventPipeProvider>()
             {
                 new EventPipeProvider("System.Runtime", EventLevel.Informational, 0, new Dictionary<string, string>() {
                     { "EventCounterIntervalSec", "1" }
@@ -96,33 +111,55 @@ namespace Microsoft.Diagnostics.NETCore.Client
             }
         }
 
+        [Fact]
+        public Task EventPipeSessionUnavailableTest()
+        {
+            return EventPipeSessionUnavailableTestCore(useAsync: false);
+        }
+
+        [Fact]
+        public Task EventPipeSessionUnavailableTestAsync()
+        {
+            return EventPipeSessionUnavailableTestCore(useAsync: true);
+        }
+
         /// <summary>
         /// Tries to start an EventPipe session on a non-existent process
         /// </summary>
-        [Fact]
-        public void EventPipeSessionUnavailableTest()
+        private async Task EventPipeSessionUnavailableTestCore(bool useAsync)
         {
             List<int> pids = new List<int>(DiagnosticsClient.GetPublishedProcesses());
             int arbitraryPid = 1;
 
-            DiagnosticsClient client = new DiagnosticsClient(arbitraryPid);
+            DiagnosticsClientApiShim clientShim = new DiagnosticsClientApiShim(new DiagnosticsClient(arbitraryPid), useAsync);
 
-            Assert.Throws<ServerNotAvailableException>(() => client.StartEventPipeSession(new List<EventPipeProvider>()
+            await Assert.ThrowsAsync<ServerNotAvailableException>(() => clientShim.StartEventPipeSession(new List<EventPipeProvider>()
             {
                 new EventPipeProvider("Microsoft-Windows-DotNETRuntime", EventLevel.Informational)
             }));
         }
 
+        [Fact]
+        public Task StartEventPipeSessionWithSingleProviderTest()
+        {
+            return StartEventPipeSessionWithSingleProviderTestCore(useAsync: false);
+        }
+
+        [Fact]
+        public Task StartEventPipeSessionWithSingleProviderTestAsync()
+        {
+            return StartEventPipeSessionWithSingleProviderTestCore(useAsync: true);
+        }
+
         /// <summary>
         /// Test for the method overload: public EventPipeSession StartEventPipeSession(EventPipeProvider provider, bool requestRundown=true, int circularBufferMB=256)
         /// </summary>
-        [Fact]
-        public void StartEventPipeSessionWithSingleProviderTest()
+        private async Task StartEventPipeSessionWithSingleProviderTestCore(bool useAsync)
         {
             using TestRunner runner = new TestRunner(CommonHelper.GetTraceePathWithArgs(), output);
             runner.Start(timeoutInMSPipeCreation: 15_000, testProcessTimeout: 60_000);
-            DiagnosticsClient client = new DiagnosticsClient(runner.Pid);
-            using (var session = client.StartEventPipeSession(new EventPipeProvider("Microsoft-Windows-DotNETRuntime", EventLevel.Informational)))
+            DiagnosticsClientApiShim clientShim = new DiagnosticsClientApiShim(new DiagnosticsClient(runner.Pid), useAsync);
+            using (var session = await clientShim.StartEventPipeSession(new EventPipeProvider("Microsoft-Windows-DotNETRuntime", EventLevel.Informational)))
             {
                 Assert.True(session.EventStream != null);
             }
