@@ -4,13 +4,9 @@
 
 using System;
 using System.Collections.Generic;
+using System.Threading.Tasks;
 using Xunit;
 using Xunit.Abstractions;
-
-using Microsoft.Diagnostics.Tracing;
-using Microsoft.Diagnostics.TestHelpers;
-using Microsoft.Diagnostics.NETCore.Client;
-using Xunit.Extensions;
 
 namespace Microsoft.Diagnostics.NETCore.Client
 {
@@ -23,11 +19,22 @@ namespace Microsoft.Diagnostics.NETCore.Client
             output = outputHelper;
         }
 
+        [Fact]
+        public Task BasicEnvTest()
+        {
+            return BasicEnvTestCore(useAsync: false);
+        }
+
+        [Fact]
+        public Task BasicEnvTestAsync()
+        {
+            return BasicEnvTestCore(useAsync: true);
+        }
+
         /// <summary>
         /// A simple test that collects process environment.
         /// </summary>
-        [Fact]
-        public void BasicEnvTest()
+        private async Task BasicEnvTestCore(bool useAsync)
         {
             // as the attribute says, this test requires 5.0-rc1 or newer.  This has been tested locally on
             // an rc1 build and passes.  It is equivalent to the dotnet/runtime version of this test.
@@ -36,8 +43,8 @@ namespace Microsoft.Diagnostics.NETCore.Client
             string testVal = "BAR";
             runner.AddEnvVar(testKey, testVal);
             runner.Start(timeoutInMSPipeCreation: 3000);
-            DiagnosticsClient client = new DiagnosticsClient(runner.Pid);
-            Dictionary<string,string> env = client.GetProcessEnvironment();
+            var clientShim = new DiagnosticsClientApiShim(new DiagnosticsClient(runner.Pid), useAsync);
+            Dictionary<string,string> env = await clientShim.GetProcessEnvironment();
 
             Assert.True(env.ContainsKey(testKey) && env[testKey].Equals(testVal));
 
